Extracted BankAccountsView from BankAccountsViewController
