+++ CybridSDK/Components/IdentityComponent/IdentityView.swift
@@ -55,5 +55,4 @@ public final class IdentityView: Component {
             }
         }
     }
-    
 }
